ugi
el ugi

+++ blog/database/migrations/2019_06_06_164001_create_trabajo_titulacions_table.php
@@ -23,7 +23,7 @@ class CreateTrabajoTitulacionsTable extends Migration
             $table->integer('id_actividad')->unsigned();
             $table->integer('id_organizacion')->unsigned()->nullable();
             $table->integer('id_academico')->unsigned();
-            $table->double('nota')->unsigned()->nullable();
+            $table->float('nota')->unsigned()->nullable();
             $table->integer('semestre')->unsigned()->nullable();
             $table->integer('numero')->unsigned()->nullable();
             $table->timestamps();
